refactor(Money) : lotto 패키지로 이동
