+++ src/test/java/lotto/domain/lotto/MoneyTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -11,6 +11,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 import lotto.domain.factory.MoneyFactory;
 import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.Money;
 
 public class MoneyTest {
